+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added Contract Details Panel](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/17)
 - [Added a new Config Context, added support for loginParams](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/14)
 - [Updated packages](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/13)